Correction #1
Correction of code.

+++ src/main/java/pl/ug/kuznia/todoappapi/internal/card/CardService.java
@@ -1,7 +1,9 @@
 package pl.ug.kuznia.todoappapi.internal.card;
 
 import org.springframework.stereotype.Service;
+import pl.ug.kuznia.todoappapi.exceptions.APIRequestException;
 
+import javax.validation.constraints.NotNull;
 import java.util.List;
 
 @Service
@@ -19,15 +21,21 @@ public class CardService {
         return cardRepository.save(cardToAdd);
     }
 
-    public void deleteCard(Long id){
-         cardRepository.deleteById(id);
+    public void deleteCard(Long id) {
+        if (cardRepository.existsById(id)) {
+            cardRepository.deleteById(id);
+        } else {
+            throw new APIRequestException("Couldn't find card id to delete");
+        }
     }
+    public Card updateCard(CardDTO cardDTO) {
+        if (cardRepository.existsById(cardDTO.getId())) {
+            Card cardToUpdate = cardDTO.updateCard();
+            return cardRepository.save(cardToUpdate);
+        } else {
+            throw new APIRequestException("Couldn't find card id to update");
+        }
 
-    public Card updateCard(Long id, String title, String description) {
-        Card cardToUpdate = cardRepository.findById(id).get();
-        cardToUpdate.setTitle(title);
-        cardToUpdate.setDescription(description);
-        return cardRepository.save(cardToUpdate);
     }
 
     public List<Card> findAllCards() {
